5/29
Reformed item editing/placement script to be much better and more flexible.

Next step is to make a script that spawns an item into the item held state without already being on the grid.

This will be used for the dev menu and spawning items in by clicking and dragging!