6/10
Added some prefabs to add to the atk list.

Need to code a new method that runs if atkCount < 5 disable the first program and lerp all the others forward one. This could be incorporated into the existing ScrollUI method somehow possibly, but it will probably be cleaner to have two separate methods.

Also, might be cleaner to just have one method that checks the atkCount and disables / enabled spriterenderers based on that.