6/11
Reworked the attackUI system to work and update accordingly with the actual list of attackPrograms.

The ui display for the deck count needs to be updated because currentAttackProgramAmount in AttackProgramData is now counting from 0 first so that the two lists work in parallel. Another value must be calulated for the UI.

To-do

Create some sort of parabolic interpolation helper function to create a cool ui effect when the programs are moved.

Need to create a sprite the directly defines the starting position so they can be passed through the unity editor rather than gotten at Start().