6/19
Hitbox and Hurtbox system seemed to be implemented good for now. Next steps are probably making a test dummy prefab so test player attacks. Before this, I should make a health bar prefab to attach to enemies, and a damage number mechanic that when attatched to a object and given the object's hurtbox, damage number float above the enemy when dealt.