6/22
Added dummy prefab logic and a prefab healthbar to attach to future enemies. Added a universal animator that is currently being used in the dummy Logic script to visual each hit on the onHit event in its respective hurtbox.

Need to add a damage numbers script still that float and fades damage numbers above an attacked entity.