6/29
Added some basic logic for hitboxes to be toggled on and off on certain frames of the animator class.

Added a Player logic script. Right now in this script, the player will return to idle after any animation complete in its animator. This isn't ideal long term because the animator is built to chain animations for certain future programs or attacks and this behavior could conflict with an ideal animation subscribed to run on any animation complete call.