7/14
Reworked the animator to function based on a totalAnimationTime

Added an animation timer that displays the progress of the current playing attack.

Should research how to implement some sort of acceleration into this bar (reference the aesprite animation concept) Not sure how this would work though as the animation progresses at a constant rate. There would have to be some separate math that took the end time and displayed the bar based on a quadratic function that ends at the same time as the animation.