7/23
Program UI system has been applied to both attacking and defending. Now, each of the two instances of Program UI must be set to Attack or Defense via a new enum. This will determine if it tracks isAttacking vs isDefending. Also, in the program class inherited to every program, whether or no the program is Attack or Defense must be specified by the same enum. The enum is declared in ProgramUI right now. If more enums get declared in the future we should probably move them all to one enum delcaration script.