7/26
Added a defense program with only animation logic at the moment. Need to add actual methods for making the player invulnerable in playerLogic script.

Need to add way for programs to check if an animation is an attack, defense, or other animation to only scroll ui when its type of animation plays. Right now, all program logic script indifferently subscibe to OnAnimationComplete which trigger both attack and defense scrolls at the same time.