11/29
Introduced the DashLogic script. Everything seems to be talking to each other correctly. Need to add the correct sprites and thresholds to the 1_Dash program asset. Need to change the math for the lerp in PlayerMovement script to be less slidey/slippery looking. Need to add some sort of tile targeting script that has a changing distance threshold depending on the current program.